quell/fix issues raised by the Intel compiler on Altix
svn:revision:26436
cvs:account:brlcad
cvs:branch:trunk

+++ src/mged/bodyio.c
@@ -249,7 +249,7 @@ cmd_export_body(ClientData clientData, Tcl_Interp *interp, int argc, char **argv
     int				fd;
     void			*bufp;
     size_t			nbytes = 0;
-    int				written;
+    long int			written;
     struct bu_external		ext;
     struct db5_raw_internal	raw;
     struct rt_db_internal	intern;
