+++ src/mged/columns.c
@@ -46,7 +46,7 @@ static const char RCSid[] = "@(#)$Header$ (BRL)";
 #include "common.h"
 
 #include <stdio.h>
-
+#include <stdlib.h>
 #ifdef HAVE_STRING_H
 #  include <string.h>
 #else
@@ -286,8 +286,8 @@ vls_long_dpp(
 	    bu_vls_spaces(vls, max_nam_len - strlen( dp->d_namep ) );
 	    bu_vls_printf(vls, " %s", type );
 	    bu_vls_spaces(vls, max_type_len - strlen( type ) );
-	    bu_vls_printf(vls,  " %2d %2d %d\n",
-		    dp->d_major_type, dp->d_minor_type, dp->d_len);
+	    bu_vls_printf(vls,  " %2d %2d %ld\n",
+		    dp->d_major_type, dp->d_minor_type, (long)dp->d_len);
     }
   }
 }
